keyring.New() -> keyring.NewLegacy()
