+++ crypto/keyring/keybase.go
@@ -77,13 +77,3 @@ type Keybase interface {
 	// SupportedAlgosLedger returns a list of signing algorithms supported by the keybase's ledger integration
 	SupportedAlgosLedger() []SigningAlgo
 }
-
-// LegacyKeybase is implemented by the legacy keybase implementation.
-type LegacyKeybase interface {
-	List() ([]Info, error)
-	Export(name string) (armor string, err error)
-	ExportPrivKey(name, decryptPassphrase, encryptPassphrase string) (armor string, err error)
-	ExportPubKey(name string) (armor string, err error)
-	Update(name, oldpass string, getNewpass func() (string, error)) error
-	Close() error
-}
